+++ Client/AnpanMMO/Source/AnpanMMO/Active/UI/Menu/Skill/SkillMenu.h
@@ -7,6 +7,46 @@
 #include "SkillMenu.generated.h"
 
 /**
+ * スキルツリーノード
+ */
+USTRUCT(BlueprintType)
+struct FSkillTreeNode
+{
+
+	GENERATED_USTRUCT_BODY()
+
+public:
+
+	// ノードＩＤ
+	uint32 NodeId;
+
+	// スキルＩＤ
+	UPROPERTY(BlueprintReadOnly)
+	int32 SkillId;
+
+	// 必要なレベル
+	UPROPERTY(BlueprintReadOnly)
+	int32 NeedLevel;
+
+	// コスト
+	UPROPERTY(BlueprintReadOnly)
+	int32 Cost;
+
+	// 開いているか？
+	UPROPERTY(BlueprintReadOnly)
+	bool bIsOpened;
+
+	// 親ノードが開いているか？
+	UPROPERTY(BlueprintReadOnly)
+	bool bIsParentOpened;
+
+	// 座標.
+	UPROPERTY(BlueprintReadOnly)
+	FVector2D Position;
+
+};
+
+/**
  * スキルメニュー
  */
 UCLASS()
